Multiple changes to lcp and grpc solver
